rename io_hdf5 to hdf5_handler
