+++ simtools/io_operations/hdf5_handler.py
